+++ examiner/__init__.py
@@ -1,5 +1,7 @@
 """
 Meta data for code
 """
+from examiner.helper_functions import check_for_tags as tags
+
 # Version structure major.minor[.patch][sub]
 __version__ = '1.4.0'