+++ test/test_exam_text_case.py
@@ -14,6 +14,7 @@ if path not in sys.path:
 
 import examiner.exceptions as exce
 from examiner.exam_test_case import ExamTestCase
+from examiner import tags
 
 class Test_ExamTestCase(unittest.TestCase):
 
@@ -197,31 +198,64 @@ class Test_ExamTestCase(unittest.TestCase):
         class Test1Tags1(ExamTestCase):
             USER_TAGS = ["dont_skip"]
 
+            @tags("skip")
             def test_a_foo(self):
-                self.tags = ["skip"]
                 self.assertEqual('correct', 'incorrect')
 
         test = Test1Tags1('test_a_foo')
 
         with self.assertRaises(SkipTest) as _:
             test.test_a_foo()
+        # check that method was decorated for tags
+        self.assertEqual(getattr(test.test_a_foo, "__wrapped__").__name__, "test_a_foo")
 
 
 
     def test_run_test_by_tags(self):
         """
-        Tests that an overwritten assert runs the test without being skipped.
+        Tests that an overwritten test runs the test without being skipped.
         """
         class Test1Tags1(ExamTestCase):
             USER_TAGS = ["dont_skip"]
 
+            @tags("dont_skip")
             def test_a_foo(self):
-                self.tags = ["dont_skip"]
                 return "Not Skipped"
 
         test = Test1Tags1('test_a_foo')
         self.assertEqual(test.test_a_foo(), "Not Skipped")
+        # check that method was decorated for tags
+        self.assertEqual(getattr(test.test_a_foo, "__wrapped__").__name__, "test_a_foo")
 
+
+    def test_skip_test_when_error(self):
+        """
+        Tests that SkipTest is raised when tested code raise an exception.
+        This was a bug before.
+        """
+        class Test1Tags1(ExamTestCase):
+            USER_TAGS = ["dont_skip"]
+
+            @tags("skip")
+            def test_a_foo(self):
+                raise KeyError()
+
+        test = Test1Tags1('test_a_foo')
+
+        with self.assertRaises(SkipTest) as _:
+            test.test_a_foo()
+
+    def test_passing_simpel_test(self):
+        """
+        Check that a normal test works
+        """
+        class Test1Simpel(ExamTestCase):
+
+            def test_a_foo(self):
+                self.assertEqual("hej", "hej")
+
+        test = Test1Simpel('test_a_foo')
+        self.assertEqual(test.USER_TAGS, [])
 
 if __name__ == '__main__':
     # runner = unittest.TextTestRunner(resultclass=ExamTestResult, verbosity=2)