+++ test/test_exam_text_case_asserts.py
@@ -5,14 +5,12 @@ import sys
 import os
 import unittest
 from unittest.runner import _WritelnDecorator
-from unittest import SkipTest
 
 proj_path = os.path.dirname(os.path.realpath(__file__ + "/../"))
 path = proj_path + "/examiner"
 if path not in sys.path:
     sys.path.insert(0, path)
 
-import exceptions as exce
 from exam_test_case import ExamTestCase
 
 class Test_ExamTestCase(unittest.TestCase):
@@ -23,7 +21,6 @@ class Test_ExamTestCase(unittest.TestCase):
         """
         class Test1AssertNotIn(ExamTestCase):
             def test_a_foo(self_):
-                self_.tags = ['test', 'assert']
                 self_.assertNotIn('correct', ['incorrect'])
 
         test = Test1AssertNotIn('test_a_foo')
@@ -31,9 +28,7 @@ class Test_ExamTestCase(unittest.TestCase):
 
         self.assertEqual(test.correct_answer, "'correct'")
         self.assertEqual(test.student_answer, "['incorrect']")
-        self.assertEqual(test.tags, ['test', 'assert'])
-        # check that method was decorated for tags
-        self.assertEqual(getattr(test.assertNotIn, "__wrapped__").__name__, "assertNotIn")
+
 
 
 
@@ -51,7 +46,6 @@ class Test_ExamTestCase(unittest.TestCase):
 
         self.assertEqual(test.correct_answer, "'correct'")
         self.assertEqual(test.student_answer, "['correct', 'incorrect']")
-        self.assertEqual(test.tags, [])