+++ examiner/__init__.py
@@ -1,7 +1,9 @@
 """
 Meta data for code
 """
-from examiner.helper_functions import check_for_tags as tags
+from examiner.exam_test_case import ExamTestCase
+from examiner.exam_test_result import ExamTestResult
+from examiner.helper_functions import check_for_tags as tags, find_path_to_assignment, import_module
 
 # Version structure major.minor[.patch][sub]
 __version__ = '1.4.0'