+++ test/test_exam_text_result.py
@@ -6,13 +6,8 @@ import os
 import unittest
 from unittest.runner import _WritelnDecorator
 
-proj_path = os.path.dirname(os.path.realpath(__file__ + "/../"))
-path = proj_path + "/examiner"
-if path not in sys.path:
-    sys.path.insert(0, path)
-
-from exam_test_result import ExamTestResult
-from exam_test_case import ExamTestCase
+from examiner import ExamTestResult
+from examiner import ExamTestCase