+++ CHANGELOG.md
@@ -14,8 +14,12 @@ Available types:
 - `Security` in case of vulnerabilities.
 
 ## [Unreleased]
+### Added
+- Added som common functions/classes can be imported directly from examiner.
+
 ### Fixed
 - --tags will not properly skip tests if they don't match.
+
 
 ## [1.4.0] - 2021-05-20
 ### Changed